Add safety to verify windows vs *nix

+++ src/clean-aws-credentials-cache/index.js
@@ -2,10 +2,11 @@ const fs = require('fs');
 
 //https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_use_revoke-sessions.html
 function cleanAwsCredentialsCache() {
-    if (fs.existsSync('~/.aws/cli/cache')) {
+    // Based on current values: https://nodejs.org/api/process.html#process_process_platform
+    if (process.platform !== "win32" && fs.existsSync('~/.aws/cli/cache')) {
         fs.rmdirSync('~/.aws/cli/cache', { recursive: true });
     }
-    if (fs.existsSync('%UserProfile%\\.aws\\cli\\cache')) {
+    if (process.platform === "win32" && fs.existsSync('%UserProfile%\\.aws\\cli\\cache')) {
         fs.rmdirSync('%UserProfile%\\.aws\\cli\\cache', { recursive: true });
     }
 }
